% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayobeamer.R
\name{mayobeamer}
\alias{mayobeamer}
\title{A Mayo themed Beamer template}
\usage{
mayobeamer(..., extra_preamble = NULL)
}
\arguments{
\item{...}{reserved for future arguments}

\item{extra_preamble}{Additional tex files to include.}
}
\description{
A Mayo themed Beamer template
}
