% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayodocx.R
\name{mayodocx}
\alias{mayodocx}
\title{A mayo themed docx format}
\usage{
mayodocx(...)
}
\arguments{
\item{...}{arguments used by \link[rmarkdown]{word_document}}
}
\description{
Format for converting from R Markdown to MS Word
based on \link[officedown]{rdocx_document}.
}
