% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayoioslides.R
\name{mayoioslides}
\alias{mayoioslides}
\title{A Mayo themed Ioslides Template}
\usage{
mayoioslides(...)
}
\arguments{
\item{...}{reserved for future arguments}
}
\description{
A Mayo themed Ioslides Template
}
