% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayomoon_reader.R
\name{mayomoon_reader}
\alias{mayomoon_reader}
\title{A mayo themed template for remark.js slides}
\usage{
mayomoon_reader(
  nature = list(countIncrementalSlides = "no", highlightLines = "yes", highlightStyle =
    "github"),
  ...
)
}
\arguments{
\item{nature}{See ?xaringan::moon_reader}

\item{...}{Additional arguments passed to xaringan::moon_reader}
}
\description{
A mayo themed template for remark.js slides
}
