% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mayoioslides.R
\name{mayoioslides}
\alias{mayoioslides}
\title{A Mayo themed Ioslides Template}
\usage{
mayoioslides(..., extra_css = NULL)
}
\arguments{
\item{...}{reserved for future arguments}

\item{extra_css}{Additional CSS files to include.}
}
\description{
A Mayo themed Ioslides Template
}
